+++ lihil/endpoint/__init__.py
@@ -3,4 +3,3 @@ from .returns import CustomEncoder as CustomEncoder
 from .signature import EndpointSignature as EndpointSignature
 from .signature import ParseResult as ParseResult
 from .signature import RequestParam as RequestParam
-from .endpoint import Endpoint